Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-generativeai==0.3.1
python-multipart
pandas
